// src/app/shared/models/sl-transportation-method.model.ts
export interface SLTransportationMethod {
    Destintation: string;

    DisplayTime: string;

    ExpectedDateTime: string;

    JourneyDirection: number;

    JourneyNumber: number;

    LineNumber: string;

    StopAreaName: string;

    StopAreaNumber: number;

    StopPointNumber: number;

    TimeTabledDateTime: string;

    TransportMode: string;
}